rename RelatedListOverContent to RelatedListOverContentB
