+++ src/views/Article.vue
@@ -452,7 +452,7 @@ import PopInAd from '../components/PopInAd.vue'
 import PopList from '../components/article/PopList.vue'
 import PopListVert from '../components/article/PopListVert.vue'
 import RelatedListInContent from '../components/article/RelatedListInContent.vue'
-import RelatedListOverContent from '../components/article/RelatedListOverContent.vue'
+import RelatedListOverContent from '../components/article/RelatedListOverContentB.vue'
 import RecommendList from '../components/article/RecommendList.vue'
 import ShareTools from '../components/article/ShareTools.vue'
 import WineWarning from '../components/WineWarning.vue'
